requirements.txt
annotated-types==0.7.0
anyio==4.5.2
awscli==1.36.33
beautifulsoup4==4.12.3
boto3==1.35.92
botocore==1.35.92
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
docutils==0.16
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
jmespath==1.0.1
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
pip==20.0.2
pkg-resources==0.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-core==2.27.2
pyOpenSSL==24.3.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.4
setuptools==44.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.33.0